+++ examples/encode_decode_on_the_fly/encode_decode_on_the_fly.cpp
@@ -3,14 +3,6 @@
 // See accompanying file LICENSE.rst or
 // http://www.steinwurf.com/licensing
 
-#include <cstdlib>
-#include <ctime>
-#include <iostream>
-#include <vector>
-#include <algorithm>
-
-#include <kodocpp/kodocpp.hpp>
-
 /// @example encode_decode_on_the_fly.cpp
 ///
 /// This example shows how to use a storage aware encoder which will
@@ -18,6 +10,15 @@
 /// specified. This can be useful in cases where the symbols that
 /// should be encoded are produced on-the-fly. The decoder will also
 /// allow you to detect whether the symbols have been partially decoded.
+
+#include <cstdint>
+#include <cstdlib>
+#include <ctime>
+#include <iostream>
+#include <vector>
+#include <algorithm>
+
+#include <kodocpp/kodocpp.hpp>
 
 int main(void)
 {
@@ -31,16 +32,16 @@ int main(void)
 
     // Initilization of encoder and decoder
     kodocpp::encoder_factory encoder_factory(
-        kodo_on_the_fly,
-        kodo_binary8,
+        kodocpp::codec::on_the_fly,
+        kodocpp::field::binary8,
         max_symbols,
         max_symbol_size);
 
     kodocpp::encoder encoder = encoder_factory.build();
 
     kodocpp::decoder_factory decoder_factory(
-        kodo_on_the_fly,
-        kodo_binary8,
+        kodocpp::codec::on_the_fly,
+        kodocpp::field::binary8,
         max_symbols,
         max_symbol_size);
 
@@ -78,8 +79,7 @@ int main(void)
             uint32_t rank = encoder.rank();
 
             // Calculate the offset to the next symbol to insert
-            uint8_t* symbol = data_in.data() + (rank * encoder.symbol_size());
-
+            uint8_t* symbol = data_in.data() + rank * encoder.symbol_size();
             encoder.set_const_symbol(rank, symbol, encoder.symbol_size());
         }
 
@@ -105,8 +105,8 @@ int main(void)
         // Check if the decoder is partially complete
         // The decoder has to support partial decoding tracker for
         // on-the-fly decoding
-        if (decoder.has_partial_decoding_tracker() &&
-            decoder.is_partial_complete())
+        if (decoder.has_partial_decoding_interface() &&
+            decoder.is_partially_complete())
         {
             for (uint32_t i = 0; i < decoder.symbols(); ++i)
             {