+++ src/kodocpp/encoder_factory.hpp
@@ -10,6 +10,8 @@
 
 #include <kodoc/kodoc.h>
 
+#include "field.hpp"
+#include "codec.hpp"
 #include "factory.hpp"
 #include "encoder.hpp"
 
@@ -23,18 +25,28 @@ namespace kodocpp
 
     public:
 
-        encoder_factory(kodo_code_type code, kodo_finite_field field,
+        /// Constructs a new encoder factory (for shallow storage encoders)
+        /// @param codec This parameter determines the encoding algorithms used.
+        /// @param field The finite field that should be used by the
+        ///        encoder.
+        /// @param max_symbols The maximum number of symbols supported by
+        ///        encoders built with this factory.
+        /// @param max_symbol_size The maximum symbol size in bytes supported
+        ///        by encoders built using the returned factory.
+        encoder_factory(codec codec, field field,
                         uint32_t max_symbols, uint32_t max_symbol_size) :
-            factory(kodo_new_encoder_factory(code, field, max_symbols,
-                    max_symbol_size), [](kodo_factory_t factory)
+            factory(kodoc_new_encoder_factory((int32_t)codec, (int32_t)field,
+                    max_symbols, max_symbol_size), [](kodoc_factory_t factory)
                     {
-                        kodo_delete_factory(factory);
+                        kodoc_delete_factory(factory);
                     })
         { }
 
+        /// Builds a new encoder with the factory using the specified parameters
+        /// @return The new encoder
         coder_type build()
         {
-            kodo_coder_t coder = kodo_factory_new_encoder(m_factory.get());
+            kodoc_coder_t coder = kodoc_factory_build_coder(m_factory.get());
             return coder_type(coder);
         }
     };