+++ src/kodocpp/kodocpp.hpp
@@ -1,10 +1,12 @@
-// Copyright Steinwurf ApS 2011-2013.
+// Copyright Steinwurf ApS 2014.
 // Distributed under the "STEINWURF RESEARCH LICENSE 1.0".
 // See accompanying file LICENSE.rst or
 // http://www.steinwurf.com/licensing
 
 #pragma once
 
+#include "field.hpp"
+#include "codec.hpp"
 #include "encoder_factory.hpp"
 #include "decoder_factory.hpp"
 #include "encoder.hpp"