+++ test/src/test_decoder.cpp
@@ -13,41 +13,49 @@
 
 namespace kodocpp
 {
-namespace
-{
-    void test_decoder(uint32_t max_symbols, uint32_t max_symbol_size,
-        kodo_code_type code_type, kodo_finite_field finite_field)
+    static void test_decoder(uint32_t max_symbols, uint32_t max_symbol_size,
+        codec code, field field)
     {
         decoder_factory decoder_factory(
-            code_type,
-            finite_field,
+            code,
+            field,
             max_symbols,
             max_symbol_size);
 
         decoder decoder = decoder_factory.build();
-        test_coder(decoder, max_symbols, max_symbol_size, code_type);
+        test_coder(decoder, max_symbols, max_symbol_size, code);
 
         // Decoder methods
-        EXPECT_GE(0U, decoder.symbols_uncoded());
-        EXPECT_GE(0U, decoder.symbols_seen());
-
-        if (code_type == kodo_on_the_fly ||
-            code_type == kodo_sliding_window)
+        // Some codecs do not provide write_payload, i.e. recoding
+        if (code == codec::seed || code == codec::sparse_seed ||
+            code == codec::fulcrum || code == codec::reed_solomon)
         {
-            EXPECT_TRUE(decoder.has_partial_decoding_tracker());
-            EXPECT_FALSE(decoder.is_partial_complete());
+            EXPECT_FALSE(decoder.has_write_payload());
+        }
+        else
+        {
+            EXPECT_TRUE(decoder.has_write_payload());
+        }
+
+        EXPECT_GE(0U, decoder.symbols_uncoded());
+        EXPECT_GE(0U, decoder.symbols_partially_decoded());
+
+        if (code == codec::on_the_fly ||
+            code == codec::sliding_window)
+        {
+            EXPECT_TRUE(decoder.has_partial_decoding_interface());
+            EXPECT_FALSE(decoder.is_partially_complete());
             for (uint32_t i = 0; i < decoder.symbols(); ++i)
             {
                 EXPECT_FALSE(decoder.is_symbol_pivot(i));
             }
         }
-        else if (code_type == kodo_full_vector ||
-                 code_type == kodo_perpetual)
+        else if (code == codec::full_vector ||
+                 code == codec::perpetual)
         {
-            EXPECT_FALSE(decoder.has_partial_decoding_tracker());
+            EXPECT_FALSE(decoder.has_partial_decoding_interface());
         }
     }
-}
 }
 
 TEST(test_decoder, invoke_api)
@@ -56,8 +64,6 @@ TEST(test_decoder, invoke_api)
 
     uint32_t max_symbols = rand_symbols();
     uint32_t max_symbol_size = rand_symbol_size();
-
-    test_combinations(test_decoder, max_symbols, max_symbol_size);
 
     test_combinations(test_decoder, max_symbols, max_symbol_size);
 }