+++ test/src/test_full_vector_codes.cpp
@@ -9,25 +9,15 @@
 
 #include "test_helper.hpp"
 
-
 TEST(test_full_vector_codes, basic_api)
 {
     using namespace kodocpp;
 
+    if (has_codec(codec::full_vector) == false)
+        return;
+
     uint32_t max_symbols = rand_symbols();
     uint32_t max_symbol_size = rand_symbol_size();
 
-    test_basic_api(kodo_full_vector, max_symbols, max_symbol_size);
-}
-
-TEST(test_full_vector_codes, sparse_api)
-{
-    using namespace kodocpp;
-
-    // Make sure that the decoding can complete with a lower density
-    uint32_t max_symbols = rand_symbols() + 10;
-    uint32_t max_symbol_size = rand_symbol_size();
-
-    test_basic_api(kodo_sparse_full_vector, kodo_full_vector, max_symbols,
-                   max_symbol_size);
+    test_basic_api(codec::full_vector, max_symbols, max_symbol_size);
 }