+++ test/src/test_helper.cpp
@@ -5,12 +5,12 @@
 
 #include <kodocpp/kodocpp.hpp>
 
-#include <gtest/gtest.h>
-
-#include <algorithm>
 #include <cstdint>
 #include <cstdlib>
+#include <algorithm>
 #include <vector>
+
+#include <gtest/gtest.h>
 
 #include "test_helper.hpp"
 
@@ -37,66 +37,68 @@ namespace kodocpp
         return rand_nonzero(elements) * granularity;
     }
 
-    void test_combinations(test_function_type test_function,
+    void test_combinations(test_function_type coder_test,
                            uint32_t max_symbols, uint32_t max_symbol_size)
     {
         SCOPED_TRACE(testing::Message() << "symbols = " << max_symbols);
         SCOPED_TRACE(testing::Message() << "symbol_size = " << max_symbol_size);
 
-        std::vector<kodo_code_type> code_types =
+        std::vector<codec> codecs =
         {
-            kodo_full_vector,
-            kodo_on_the_fly,
-            kodo_sliding_window,
-            kodo_seed,
-            kodo_sparse_seed,
-            kodo_perpetual,
-            kodo_fulcrum
+            codec::full_vector,
+            codec::on_the_fly,
+            codec::sliding_window,
+            codec::sparse_full_vector,
+            codec::seed,
+            codec::sparse_seed,
+            codec::perpetual,
+            codec::fulcrum,
+            codec::reed_solomon
         };
 
-        std::vector<std::string> code_type_names =
+        std::vector<std::string> codec_names =
         {
-            "kodo_full_vector",
-            "kodo_on_the_fly",
-            "kodo_sliding_window",
-            "kodo_seed",
-            "kodo_sparse_seed",
-            "kodo_perpetual",
-            "kodo_fulcrum"
+            "full_vector",
+            "on_the_fly",
+            "sliding_window",
+            "sparse_full_vector",
+            "seed",
+            "sparse_seed",
+            "perpetual",
+            "fulcrum",
+            "reed_solomon"
         };
 
-        std::vector<kodo_finite_field> fields =
+        for (uint32_t i = 0; i < codecs.size(); ++i)
         {
-            kodo_binary,
-            kodo_binary4,
-            kodo_binary8
-        };
+            SCOPED_TRACE(testing::Message() << "codec = " << codec_names[i]);
 
-        std::vector<std::string> field_names =
-        {
-            "kodo_binary",
-            "kodo_binary4",
-            "kodo_binary8"
-        };
+            codec code = codecs[i];
 
-        for (uint32_t i = 0; i < code_types.size(); ++i)
-        {
-            SCOPED_TRACE(testing::Message() << "code_type = "
-                                            << code_type_names[i]);
+            if (has_codec(code) == false)
+                continue;
 
-            for (uint32_t j = 0; j < fields.size(); ++j)
+            if (code != codec::reed_solomon)
             {
-                SCOPED_TRACE(testing::Message() << "field = "
-                                                << field_names[j]);
+                SCOPED_TRACE(testing::Message() << "field = binary");
+                coder_test(max_symbols, max_symbol_size, code, field::binary);
+            }
 
-                test_function(max_symbols, max_symbol_size, code_types[i],
-                              fields[j]);
+            if (code != codec::reed_solomon)
+            {
+                SCOPED_TRACE(testing::Message() << "field = binary4");
+                coder_test(max_symbols, max_symbol_size, code, field::binary4);
+            }
+
+            {
+                SCOPED_TRACE(testing::Message() << "field = binary8");
+                coder_test(max_symbols, max_symbol_size, code, field::binary8);
             }
         }
     }
 
     void test_coder(coder& coder, uint32_t symbols, uint32_t symbol_size,
-                    kodo_code_type code_type)
+                    codec code)
     {
         EXPECT_EQ(symbols, coder.symbols());
         EXPECT_EQ(symbol_size, coder.symbol_size());
@@ -104,7 +106,7 @@ namespace kodocpp
         EXPECT_GT(coder.payload_size(), symbol_size);
         EXPECT_EQ(0U, coder.rank());
 
-        if (code_type == kodo_sliding_window)
+        if (code == codec::sliding_window)
         {
             EXPECT_TRUE(coder.has_feedback_size());
             EXPECT_GT(coder.feedback_size(), 0U);
@@ -114,23 +116,20 @@ namespace kodocpp
             EXPECT_FALSE(coder.has_feedback_size());
         }
 
-        EXPECT_TRUE(coder.has_set_trace_stdout());
-        EXPECT_TRUE(coder.has_set_trace_callback());
-        EXPECT_TRUE(coder.has_set_trace_off());
-
+        EXPECT_TRUE(coder.has_trace_interface());
         coder.set_trace_stdout();
         coder.set_trace_off();
+        coder.set_zone_prefix("prefix");
     }
 
     /// Create and configure encoder instance
-    encoder create_encoder(uint32_t max_symbols, uint32_t max_symbol_size,
-                           kodo_code_type code_type,
-                           kodo_finite_field finite_field,
+    encoder create_encoder(codec code, field finite_field,
+                           uint32_t max_symbols, uint32_t max_symbol_size,
                            uint32_t& trace_counter)
     {
         // Initilization of encoder and decoder
-        encoder_factory encoder_factory(code_type, finite_field, max_symbols,
-            max_symbol_size);
+        encoder_factory encoder_factory(
+            code, finite_field, max_symbols, max_symbol_size);
 
         encoder encoder = encoder_factory.build();
 
@@ -143,7 +142,7 @@ namespace kodocpp
 
         EXPECT_GE(encoder_factory.max_payload_size(), encoder.payload_size());
 
-        // Install a custom trace function for the encoder and decoder
+        // Install a custom trace function for the encoder
         auto callback = [](uint32_t& count, const std::string& zone,
                            const std::string& data)
         {
@@ -154,12 +153,12 @@ namespace kodocpp
 
         using namespace std::placeholders;
 
-        EXPECT_TRUE(encoder.has_set_trace_callback());
+        EXPECT_TRUE(encoder.has_trace_interface());
         encoder.set_trace_callback(
             std::bind<void>(callback, std::ref(trace_counter), _1, _2));
 
-        if (code_type == kodo_sparse_full_vector ||
-            code_type == kodo_sparse_seed)
+        if (code == codec::sparse_full_vector ||
+            code == codec::sparse_seed)
         {
             // Set the coding vector density on the encoder
             encoder.set_density(0.2);
@@ -167,7 +166,7 @@ namespace kodocpp
         }
 
         // Test perpetual specific functions
-        if (code_type == kodo_perpetual)
+        if (code == codec::perpetual)
         {
             EXPECT_FALSE(encoder.pseudo_systematic());
             bool pseudo_systematic = true;
@@ -193,13 +192,12 @@ namespace kodocpp
     }
 
     /// Create and configure decoder instance
-    decoder create_decoder(uint32_t max_symbols, uint32_t max_symbol_size,
-                           kodo_code_type code_type,
-                           kodo_finite_field finite_field,
+    decoder create_decoder(codec code, field finite_field,
+                           uint32_t max_symbols, uint32_t max_symbol_size,
                            uint32_t& trace_counter)
     {
-        decoder_factory decoder_factory(code_type, finite_field, max_symbols,
-            max_symbol_size);
+        decoder_factory decoder_factory(
+            code, finite_field, max_symbols, max_symbol_size);
 
         decoder decoder = decoder_factory.build();
 
@@ -210,7 +208,7 @@ namespace kodocpp
 
         EXPECT_GE(decoder_factory.max_payload_size(), decoder.payload_size());
 
-        // Install a custom trace function for the encoder and decoder
+        // Install a custom trace function for the decoder
         auto callback = [](uint32_t& count, const std::string& zone,
                            const std::string& data)
         {
@@ -221,17 +219,15 @@ namespace kodocpp
 
         using namespace std::placeholders;
 
-        EXPECT_TRUE(decoder.has_set_trace_callback());
+        EXPECT_TRUE(decoder.has_trace_interface());
         decoder.set_trace_callback(
             std::bind<void>(callback, std::ref(trace_counter), _1, _2));
 
         return decoder;
     }
 
-    void run_test_basic_api(
-        kodo_code_type encoder_type, kodo_code_type decoder_type,
-        kodo_finite_field finite_field,
-        uint32_t max_symbols, uint32_t max_symbol_size)
+    void run_test_basic_api(codec code, field finite_field,
+                            uint32_t max_symbols, uint32_t max_symbol_size)
     {
         uint32_t encoder_trace_count = 0;
         uint32_t decoder_trace_count = 0;
@@ -245,13 +241,11 @@ namespace kodocpp
         // functions to verify correct operation with copied instances.
         // The corresponding factories are also destroyed before the coder
         // instances are used.
-        encoder = create_encoder(max_symbols, max_symbol_size,
-                                 encoder_type, finite_field,
-                                 encoder_trace_count);
+        encoder = create_encoder(code, finite_field, max_symbols,
+                                 max_symbol_size, encoder_trace_count);
 
-        decoder = create_decoder(max_symbols, max_symbol_size,
-                                 decoder_type, finite_field,
-                                 decoder_trace_count);
+        decoder = create_decoder(code, finite_field, max_symbols,
+                                 max_symbol_size, decoder_trace_count);
 
         EXPECT_EQ(encoder.payload_size(), decoder.payload_size());
 
@@ -273,14 +267,14 @@ namespace kodocpp
 
         // Assign the data buffer to the encoder so that we may start
         // to produce encoded symbols from it
-        if (encoder_type != kodo_on_the_fly)
+        if (code != codec::on_the_fly)
         {
             encoder.set_const_symbols(data_in.data(), encoder.block_size());
         }
 
         // Test feedback if we are testing sliding window
         std::vector<uint8_t> feedback;
-        if (encoder_type == kodo_sliding_window)
+        if (code == codec::sliding_window)
         {
             uint32_t feedback_size = 0;
 
@@ -296,7 +290,7 @@ namespace kodocpp
         {
             EXPECT_GE(encoder.rank(), decoder.rank());
 
-            if (encoder_type == kodo_on_the_fly)
+            if (code == codec::on_the_fly)
             {
                 // Randomly choose to add a new symbol (with 50% probability)
                 // if the encoder rank is less than the maximum number of
@@ -323,7 +317,7 @@ namespace kodocpp
             // Pass that packet to the decoder
             decoder.read_payload(payload.data());
 
-            if (encoder_type == kodo_sliding_window)
+            if (code == codec::sliding_window)
             {
                 decoder.write_feedback(feedback.data());
                 encoder.read_feedback(feedback.data());
@@ -332,26 +326,33 @@ namespace kodocpp
             // Check the decoder rank and symbol counters
             EXPECT_GE(encoder.rank(), decoder.rank());
             EXPECT_GE(decoder.rank(), decoder.symbols_uncoded());
-            EXPECT_GE(decoder.rank(), decoder.symbols_seen());
+            EXPECT_GE(decoder.rank(), decoder.symbols_partially_decoded());
+            EXPECT_EQ(decoder.symbols() - decoder.rank(),
+                      decoder.symbols_missing());
 
             // Check if the decoder is partially complete
             // The decoder has to support partial decoding tracker for
             // on-the-fly decoding
-            if (decoder.has_partial_decoding_tracker() &&
-                decoder.is_partial_complete())
+            if (decoder.has_partial_decoding_interface() &&
+                decoder.is_partially_complete())
             {
                 for (uint32_t i = 0; i < decoder.symbols(); ++i)
                 {
+                    // Go through all symbols that are already decoded
                     if (decoder.is_symbol_uncoded(i))
                     {
                         // All uncoded symbols must have a pivot
                         EXPECT_TRUE(decoder.is_symbol_pivot(i));
+                        // The uncoded symbols cannot be missing
+                        EXPECT_FALSE(decoder.is_symbol_missing(i));
+                        // The uncoded symbols cannot be partially decoded
+                        EXPECT_FALSE(decoder.is_symbol_partially_decoded(i));
 
                         uint32_t offset = i * max_symbol_size;
                         uint8_t* original = data_in.data() + offset;
 
-                        // Get the decoded symbol and verify it against the
-                        // original data
+                        // Get the decoded symbol and compare it to the
+                        // original symbol
                         uint8_t* decoded_symbol = data_out.data() + offset;
                         EXPECT_EQ(memcmp(
                             original, decoded_symbol, max_symbol_size), 0);
@@ -368,35 +369,30 @@ namespace kodocpp
         EXPECT_GT(decoder_trace_count, 0U);
     }
 
-    void test_basic_api(kodo_code_type encoder_type,
-                        kodo_code_type decoder_type,
-                        uint32_t symbols, uint32_t symbol_size)
+    void test_basic_api(codec code, uint32_t symbols,
+                        uint32_t symbol_size)
     {
         SCOPED_TRACE(testing::Message() << "symbols = " << symbols);
         SCOPED_TRACE(testing::Message() << "symbol_size = " << symbol_size);
 
+        if (code != codec::reed_solomon)
         {
             SCOPED_TRACE(testing::Message() << "field = binary");
-            run_test_basic_api(encoder_type, decoder_type, kodo_binary,
+            run_test_basic_api(code, field::binary,
                                symbols, symbol_size);
         }
 
+        if (code != codec::reed_solomon)
         {
             SCOPED_TRACE(testing::Message() << "field = binary4");
-            run_test_basic_api(encoder_type, decoder_type, kodo_binary4,
+            run_test_basic_api(code, field::binary4,
                                symbols, symbol_size);
         }
 
         {
             SCOPED_TRACE(testing::Message() << "field = binary8");
-            run_test_basic_api(encoder_type, decoder_type, kodo_binary8,
+            run_test_basic_api(code, field::binary8,
                                symbols, symbol_size);
         }
-    }
-
-    void test_basic_api(kodo_code_type coder_type, uint32_t symbols,
-                        uint32_t symbol_size)
-    {
-        test_basic_api(coder_type, coder_type, symbols, symbol_size);
     }
 }