+++ test/src/test_on_the_fly_codes.cpp
@@ -13,8 +13,11 @@ TEST(test_on_the_fly_codes, invoke_api)
 {
     using namespace kodocpp;
 
+    if (has_codec(codec::on_the_fly) == false)
+        return;
+
     uint32_t max_symbols = rand_symbols();
     uint32_t max_symbol_size = rand_symbol_size();
 
-   test_basic_api(kodo_on_the_fly, max_symbols, max_symbol_size);
+   test_basic_api(codec::on_the_fly, max_symbols, max_symbol_size);
 }